feat: hide edit/delete button & function to ensure only author can do those actions

+++ controllers/reviewController.js
@@ -8,7 +8,11 @@ const reviewController = {
     const campground = await Campground.findById(campgroundId)
     if (!campground) throw new ExpressError("Campground doesn't exist", 500)
 
-    const review = new Review(req.body.review)
+    const review = new Review({
+      ...req.body.review,
+      author: req.user._id,
+      createdAt: Date.now()
+    })
     campground.reviews.push(review)
     review.campground = campground
     await review.save()
@@ -18,11 +22,17 @@ const reviewController = {
   }),
   deleteReview: catchAsync(async (req, res, next) => {
     const { campId, reviewId } = req.params
-    await Campground.findByIdAndUpdate(campId, { $pull: { reviews: reviewId } })
-    if (!campground) throw new ExpressError("Campground doesn't exist", 500)
-    await Review.findByIdAndDelete(reviewId)
-    req.flash('success', 'Successfully deleted a review')
-    res.redirect(`/campgrounds/${campId}`)
+    const review = await Review.findById(reviewId).populate('author')
+    if (review.author.equals(req.user._id)) {
+      const campground = await Campground.findByIdAndUpdate(campId, { $pull: { reviews: reviewId } })
+      if (!campground) throw new ExpressError("Campground doesn't exist", 500)
+      await Review.findByIdAndDelete(reviewId)
+      req.flash('success', 'Successfully deleted a review')
+      res.redirect(`/campgrounds/${campId}`)
+    } else {
+      req.flash('error', "You can't delete this review")
+      res.redirect(`/campgrounds/${campId}`)
+    }
   })
 }
